sdk: c_api: Run sdk-example on fvp
```sh
cd sdk && make fvp

// Run on fvp
/shared # LD_LIBRARY_PATH=./ ./sdk-example-c
Getting attestation operation on aarch64.
[84954.331748] rsi: device rsi open
[84954.344525] rsi: ioctl: attestation_token
[84954.631436] rsi: device rsi released
Claims { realm_sig: Claim { label: 0, title: "Realm Signature"
28, 32, 167, 139, 197, 204, 2, 182, 242, 71, 193, 32, 23, 49,
43, 139]) }, realm_tok: [Claim { label: 10, title: "Realm chal
, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]) }, Claim { lab
...
 22, 21, 20, 19, 18, 17, 16, 31, 30, 29, 28, 27, 26, 25, 24]),
Claim[User data]: User Custom data
Claim[Platform  profile]: http://arm.com/CCA-SSD/1.0.0
Attestation test succeeded.
```
