+++ sdk/.cargo/config
@@ -3,3 +3,4 @@ target = "x86_64-unknown-linux-gnu"
 
 [target.aarch64-unknown-linux-gnu]
 linker = "../assets/toolchain/aarch64-none-linux-gnu/bin/aarch64-none-linux-gnu-gcc"
+rustflags = [ "-C", "target-feature=+crt-static" ]
